requirements.txt
# Python dependencies for Stock Market Pipeline
# Core dependencies
yfinance==0.2.33
pandas==2.1.4
numpy==1.26.2
sqlalchemy>=1.4.28,<2.0
psycopg2-binary==2.9.9

# Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0

# dbt (compatible with SQLAlchemy 1.4)
dbt-core==1.7.4
dbt-postgres==1.7.4

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
